[TI Clang Migration][Phase 3]
[TI Clang Migration][FATFS]Porting to Clang


[TI Clang Migration][PDK Utils][FATFS]Disabled PDK Utils based libs/apps

- disabling this as entry/exit hook arguments are not supported by clang


[TI Clang Migration][FreeRTOS]Disabled C99 build as it does not supprot 'asm()'


[TI Clang Migration][OSAL]Porting to Clang


[TI Clang Migration][PDK Utils][OSAL]Disabled PDK Utils based libs/apps

- disabling this as entry/exit hook arguments are not supported by clang
- disabling mcu build for TIRTOS


[TI Clang Migration][Transport]Porting to Clang


[TI Clang Migration][PDK Utils][Transport]Disabled PDK Utils based libs/apps
- Disabled NIMU as it is only built for TIRTOS


[TI Clang Migration][PDK Utils]Disabled PDK Utils based libs/apps
- disabling this as entry/exit hook arguments are not supported by clang


[TI Clang Migration]Porting to Clang

- Added '*.d' and '*.o' for ignoring these compiler generated files


[TI Clang Migration][Sci-client]Porting to Clang


[Keywriter]Disabled application for TI Clang Migration, for build fix


[TI Clang Migration][Keywriter]


[TI Clang Migration][SBL]Porting to Clang patch-3

- Fixed mcu1_0, mcu1_1 builds


[TI Clang Migration][SBL]Porting to Clang patch-4

- Fixed mcu1_0, mcu1_1 builds: increased MSMC3 memory size


[TI Clang Migration][Sci-client]Porting to Clang

+++ packages/ti/boot/sbl/example/k3MulticoreApp/sbl_smp_multicore.c
@@ -178,7 +178,7 @@ int main()
         /* Try booting all cores */
         if (proc_id_list[core_id] != 0xBAD00000)
         {
-            SBL_SlaveCoreBoot(core_id, NULL, &sblSmpTestEntry, SBL_REQUEST_CORE);
+            SBL_SlaveCoreBoot(core_id, (uint32_t)NULL, &sblSmpTestEntry, SBL_REQUEST_CORE);
             UART_printf("No of Cortex-A core(s) running: ");
             CSL_armR5CacheInv((const void *)MPU_POKE_MEM_ADDR, sizeof (int));
             UART_printf("%d\r\n", *(MPU_POKE_MEM_ADDR));
@@ -241,7 +241,7 @@ int main()
             memset(((void *)(SblBtcmAddr[core_id - MCU1_CPU0_ID])), 0xFF, 0x8000);
 
             /* Restart the core */
-            SBL_SlaveCoreBoot(core_id, NULL, &sblSmpTestEntry, SBL_REQUEST_CORE);
+            SBL_SlaveCoreBoot(core_id, (uint32_t)NULL, &sblSmpTestEntry, SBL_REQUEST_CORE);
             UART_printf("No of Cortex-R core(s) running: ");
             CSL_armR5CacheInv((const void *)MPU_POKE_MEM_ADDR, sizeof (int));
             UART_printf("%d\r\n", *(MPU_POKE_MEM_ADDR));
